requirements.txt
attrs==23.1.0
beautifulsoup4==4.12.2
biopython==1.81
certifi==2023.7.22
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
exceptiongroup==1.1.3
fonttools==4.44.3
h11==0.14.0
idna==3.4
importlib-resources==6.1.1
joblib==1.3.2
kiwisolver==1.4.5
matplotlib==3.8.2
mmtf-python==1.1.2
msgpack==1.0.3
numpy==1.25.2
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
Pmw==2.0.1+3
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
rdkit==2023.9.1
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.3
seaborn==0.13.1
selenium==4.15.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
threadpoolctl==3.2.0
tqdm==4.66.1
trio==0.23.1
trio-websocket==0.11.1
tzdata==2023.3
urllib3==2.0.7
webdriver-manager==4.0.1
wsproto==1.2.0
zipp==3.17.0
